backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
celery
redis
google-genai
python-dotenv
supabase
